Add files via upload
Project Description:
During the capstone project of Target Market Analysis in 2023, I undertook the task of designing and implementing a predictive model using Python and advanced machine learning algorithms. The objective was to identify potential customers who were likely to make a purchase in response to a specific advertisement. 

To achieve this objective, I excluded the column "Cust_Last_Purchase" from the model as it related to the target variable "Potential_Customer." By leveraging my expertise in data analysis, predictive modeling, and machine learning techniques, I developed a highly accurate model that significantly increased revenue compared to the default solution. 

The success of the project was attributed to the targeted marketing strategies that were developed based on the insights provided by the predictive model. By only sending advertisements to customers predicted as potential buyers (Class 1), we optimized resource allocation and minimized costs. 

The value and revenue of the model were calculated based on the following factors: the number of predicted positive customers, the average purchase amount of customers in the training set, and the cost of sending advertisements. This approach allowed for a thorough evaluation of the model's effectiveness and profitability.

Comparing the revenue of the best model with the revenue of the default solution, which involved sending advertisements to all customers in the test set, provided a clear indication of the superior performance and financial benefits of the optimized model.

Furthermore, a scenario was presented where a similar group of 30,000 customers would be targeted in the future, with an assumed purchase rate of 10%. By considering the Precision and Recall metrics of Class 1, the revenue of the model for this new customer set was calculated. This revenue was then compared to the revenue of the default model, which involved sending advertisements to all customers.

Through this project, I demonstrated my ability to deliver impactful, data-driven solutions that optimize customer identification and drive revenue growth. By leveraging advanced analytics and machine learning techniques, I successfully designed a predictive model that outperformed the default solution in terms of revenue generation.